LeagueHeaderView and LeagueEventView revert, refactor of LeagueHeaderCell and LeagueEventCell
