+++ Sofascore2025/Sofascore2025/Views/Events/LeagueEvent/LeagueHeaderView.swift
@@ -1,16 +1,15 @@
 //
-//  LeagueEventHeaderView.swift
+//  LeagueHeaderView.swift
 //  Sofascore2025
 //
-//  Created by Mario Brezovečki on 19.03.2025..
+//  Created by Mario Brezovečki on 11.03.2025..
 //
 
-import UIKit
 import SnapKit
+import SofaAcademic
+import UIKit
 
-class LeagueEventHeaderView: UICollectionReusableView {
-
-    static let reuseIdentifier = "LeagueEventHeaderView"
+class LeagueHeaderView: BaseView {
 
     static var height: CGFloat {
         (2 * Padding.vertical) + Constants.leagueLogoSize
@@ -34,28 +33,22 @@ class LeagueEventHeaderView: UICollectionReusableView {
     private let leagueCountryLabel = UILabel()
     private let leagueNameLabel = UILabel()
 
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        backgroundColor = .surface1
-
-        addViews()
-        styleViews()
-        setupConstraints()
+    var viewModel: LeagueViewModel? {
+        didSet {
+            if let league = viewModel {
+                setupUI(league)
+            }
+        }
     }
 
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func addViews() {
+    override func addViews() {
         addSubview(leagueLogoImageView)
         addSubview(leagueCountryLabel)
         addSubview(pointerRightIcon)
         addSubview(leagueNameLabel)
     }
 
-    func styleViews() {
+    override func styleViews() {
         leagueLogoImageView.contentMode = .scaleAspectFit
 
         leagueCountryLabel.font = .headline
@@ -70,7 +63,7 @@ class LeagueEventHeaderView: UICollectionReusableView {
         leagueNameLabel.numberOfLines = 2
     }
 
-    func setupConstraints() {
+    override func setupConstraints() {
         leagueLogoImageView.snp.makeConstraints {
             $0.size.equalTo(Constants.leagueLogoSize)
             $0.leading.equalToSuperview().inset(Padding.horizontal)
@@ -95,7 +88,7 @@ class LeagueEventHeaderView: UICollectionReusableView {
         }
     }
 
-    func bind(_ league: LeagueViewModel) {
+    private func setupUI(_ league: LeagueViewModel) {
         leagueNameLabel.text = league.name
         leagueCountryLabel.text = league.country?.name
